% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{BenchmarkConsistencyPlot}
\alias{BenchmarkConsistencyPlot}
\title{Plot median, quartile, and outlier elevations for each benchmark that is not assigned a given elevation (e.g., Benchmark 1)}
\usage{
BenchmarkConsistencyPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot median, quartile, and outlier elevations for each benchmark that is not assigned a given elevation (e.g., Benchmark 1)
}
\examples{
\dontrun{
    BenchmarkConsistencyPlot()
    BenchmarkConsistencyPlot(site = "GRBA_L_BAKR0")
}
}
