% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-qc.R
\name{BMIFunctionalMetricsPlot}
\alias{BMIFunctionalMetricsPlot}
\title{Plot functional feeding group-related richness and abundance metrics for each BMI sample.}
\usage{
BMIFunctionalMetricsPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot functional feeding group-related richness and abundance metrics for each BMI sample.
}
