% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-qc.R
\name{BMIGeneralMetricsPlot}
\alias{BMIGeneralMetricsPlot}
\title{Plot overall richness and abundance metrics for each BMI sample.}
\usage{
BMIGeneralMetricsPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot overall richness and abundance metrics for each BMI sample.
}
