% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-qc.R
\name{BMIMetricsLong}
\alias{BMIMetricsLong}
\title{Return BMI metrics data}
\usage{
BMIMetricsLong(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR1".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Return BMI metrics data
}
\examples{
\dontrun{
  BMIMetricsLong()
  BMIMetricsLong(site = c("GRBA_S_MILL1", "GRBA_S_Pine1"), field.season = "2015")
}
}
