% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-qc.R
\name{ChannelFlow}
\alias{ChannelFlow}
\title{Rank channel flow type by count for each BMI sample}
\usage{
ChannelFlow(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Rank channel flow type by count for each BMI sample
}
\examples{
\dontrun{
  ChannelFlow()
  ChannelFlow(site = "GRBA_S_STRW2", field.season = "2016")
}
}
