% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{ChemANC}
\alias{ChemANC}
\title{Calculate acid neutralizing capacity (ANC) from alkalinity (ALK2)}
\usage{
ChemANC(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, FieldSeason, SampleFrame, VisitDate, VisitType, SampleCollectionMethod, Characteristic, CharacteristicLabel, LabValue, ReportingGroup, SampleType, Flag, FlagNote, DPL, Unit
}
\description{
Calculate acid neutralizing capacity (ANC) from alkalinity (ALK2)
}
\examples{
\dontrun{
ChemANC()
ChemANC(site = "GRBA_L_DEAD0", field.season = "2018")
}
}
