% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{ChemLakeANCPlot}
\alias{ChemLakeANCPlot}
\title{Plot acid neutralizing capacity (ANC) at lakes, and include EPA thresholds}
\usage{
ChemLakeANCPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot acid neutralizing capacity (ANC) at lakes, and include EPA thresholds
}
\examples{
\dontrun{
ChemLakeANCPlot()
ChemLakeANCPlot(site = "GRBA_L_DEAD0")
}

}
