% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{CloseDatabaseConnection}
\alias{CloseDatabaseConnection}
\title{Close a connection to the Streams and Lakes Database}
\usage{
CloseDatabaseConnection(conn)
}
\arguments{
\item{conn}{A database connection pool object generated from a call to \code{OpenDatabaseConnection()}}
}
\value{
None.
}
\description{
Close a connection to the Streams and Lakes Database
}
\examples{
conn <- OpenDatabaseConnection()
CloseDatabaseConnection(conn)
}
