% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fetchAndWrangleBMI}
\alias{fetchAndWrangleBMI}
\title{Fetch BMI data from AGOL and do preliminary data wrangling}
\usage{
fetchAndWrangleBMI(
  bmi_url =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_HYDRO_BMI_Database/FeatureServer",
  agol_username = "mojn_data"
)
}
\arguments{
\item{bmi_url}{URL to AGOL BMI database}

\item{agol_username}{Authentication token (not needed for public layers)}
}
\value{
A list of data frames and metadata
}
\description{
Fetch BMI data from AGOL and do preliminary data wrangling
}
