% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetAGOLColSpec}
\alias{GetAGOLColSpec}
\title{Get column specifications for AGOL database}
\usage{
GetAGOLColSpec()
}
\value{
A list of column specifications for each table of Aquarius data.
}
\description{
Get column specifications for AGOL database
}
