% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetAquariusColSpec}
\alias{GetAquariusColSpec}
\title{Get column specifications for Aquarius data that have been written to csv.}
\usage{
GetAquariusColSpec()
}
\value{
A list of column specifications for each csv of Aquarius data.
}
\description{
Get column specifications for Aquarius data that have been written to csv.
}
