% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetColSpec}
\alias{GetColSpec}
\title{Get column specifications}
\usage{
GetColSpec()
}
\value{
A list of column specifications for each table of data.
}
\description{
Get column specifications
}
\examples{
col.spec <- GetColSpec()

# Get the names of all data tables:
data.names <- names(GetColSpec())
}
