% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{getMDLLookup}
\alias{getMDLLookup}
\title{Create tibble with MDL and ML values for each characteristic.}
\usage{
getMDLLookup()
}
\value{
A tibble with columns Characteristic, Unit, StartYear, EndYear, MDL, ML.
}
\description{
Create tibble with MDL and ML values for each characteristic.
}
