% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{LakeQcWqCleaned}
\alias{LakeQcWqCleaned}
\title{Intermediate step used to clean lake water quality data for stats and plotting functions.}
\usage{
LakeQcWqCleaned(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database",
  wq.type
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "MOJA".}

\item{site}{Optional. Site code to filter on, e.g. "LAKE_P_HOR0042".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live desert springs database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{wq.type}{Either "stream" or "lake". Indicates whether to use stream or lake water quality data.}
}
\value{
A tibble with columns for Park, FieldSeason, SiteCode, VisitDate, MeasurementDepth_m (lake only), Parameter, Units, Median, Flag, and FlagNote.
}
\description{
Limit data to primary visits and exclude data with "W" and "C" flags. Omit DO <110% or <12 mg/L.
}
