% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{LakeSurfaceElevation}
\alias{LakeSurfaceElevation}
\title{Calculates lake level elevations}
\usage{
LakeSurfaceElevation(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, VisitDate, FieldSeason,VisitType, DPL, SurveyType, BenchmarkUsed, ClosureError_ft, FinalElevation_ft.
}
\description{
Calculates lake level elevations
}
\examples{
\dontrun{
    LakeSurfaceElevation()
    LakeSurfaceElevation(site = "GRBA_L_BAKR0", field.season = "2019")
}
}
