% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{LakeWqMedian}
\alias{LakeWqMedian}
\title{Calculate median values for each water quality parameter for each lake visit.}
\usage{
LakeWqMedian(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_DEAD0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Calculate median values for each water quality parameter for each lake visit.
}
