% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{LoadStreamsAndLakes}
\alias{LoadStreamsAndLakes}
\title{Load raw data into package environment}
\usage{
LoadStreamsAndLakes(
  data_path = c("database", "aquarius", bmi_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_HYDRO_BMI_Database/FeatureServer",
    calibration_db =
    "https://services1.arcgis.com/fBc8EJBxQRMcHlei/arcgis/rest/services/MOJN_Calibration_Database/FeatureServer"),
  use_default_sql = TRUE,
  sql_drv = odbc::odbc(),
  agol_username = "mojn_data",
  agol_password =
    rstudioapi::askForPassword(paste("Please enter the password for AGOL account",
    agol_username)),
  ...
)
}
\arguments{
\item{data_path}{A path or URL to the data. Accepted inputs:
\itemize{
\item 2 URLs to the AGOL feature services containing the data (bmi_db and calibration_db)
\item a folder containing the data in csv format
\item a .zip file containing the data in csv format
\item \code{"database"} (connect to the SQL server database)
\item \code{"aquarius"} (connect to the Aquarius database)
}}

\item{use_default_sql}{Use default SQL database? Ignored if \code{data_path != "database"}.}

\item{sql_drv}{Driver to use to connect to database. Ignored if \code{data_path != "database"}.}

\item{...}{}
}
\value{
Invisibly return a list containing all raw data
}
\description{
Run this function before you do anything else.
}
\examples{
\dontrun{
LoadStreamsAndLakes()  # Read from all sources (e.g., AGOL, SQL, Aquarius, CSVs)
LoadStreamsAndLakes("aquarius")  # Read from Aquarius database only
LoadStreamsAndLakes("path/to/csv/folder")  # Read from folder of CSVs
LoadStreamsAndLakes("path/to/zipped/csvs.zip")  # Read from zip file of CSVs
}
}
