% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{MaxDQF}
\alias{MaxDQF}
\title{"Maximum" data quality flag}
\usage{
MaxDQF(flags)
}
\arguments{
\item{flags}{A character vector of data quality flag codes (C, W, I, NF)}
}
\value{
The most severe code present
}
\description{
Helper function for summarizing data with data quality flags. Given a vector of data quality flag codes, returns the "maximum," i.e. the most severe flag.
}
