% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{PlotLakeSurfaceElevation}
\alias{PlotLakeSurfaceElevation}
\title{Plot lake surface elevations over time}
\usage{
PlotLakeSurfaceElevation(
  park,
  site,
  field.season,
  include.title = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{include.title}{Include plot title? Defaults to true.}

\item{plotly}{Return an interactive plotly object instead of a ggplot object? Defaults to false.}
}
\value{
A ggplot or plotly object
}
\description{
Plot lake surface elevations over time
}
\examples{
\dontrun{
PlotLakeSurfaceElevation()
PlotLakeSurfaceElevation(site = "GRBA_L_DEAD0", plotly = TRUE)
}
}
