% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{PlotStringComparisons}
\alias{PlotStringComparisons}
\title{Plot lake levels determined by all benchmarks where string method was used}
\usage{
PlotStringComparisons(park, site, field.season)
}
\arguments{
\item{field.season}{}
}
\value{
A ggplot object
}
\description{
Plot lake levels determined by all benchmarks where string method was used
}
\examples{
\dontrun{
PlotStringComparisons()
PlotStringComparisons(site = "GRBA_L_DEAD0")
}
}
