% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmi-qc.R
\name{qcBMIDiscrepancies}
\alias{qcBMIDiscrepancies}
\title{BROKEN! FIX! Check for discrepancies between taxa count and abundance}
\usage{
qcBMIDiscrepancies(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
BROKEN! FIX! Check for discrepancies between taxa count and abundance
}
\examples{
\dontrun{
  qcBMIDiscrepancies()
  qcBMIDiscrepancies(site = "GRBA_S_MILL1")
  qcBMIDiscrepancies(site = c("GRBA_S_BAKR2", "GRBA_S_BAKR3"), field.season = "2015")
}
}
