% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{qcChemFieldDupes}
\alias{qcChemFieldDupes}
\title{Calculate the relative percent difference (RPD) for field duplicates, flag results that exceed the 30\% MQO threshold, and list all RPD values and flags.}
\usage{
qcChemFieldDupes(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Calculate the relative percent difference (RPD) for field duplicates, flag results that exceed the 30\% MQO threshold, and list all RPD values and flags.
}
\examples{
\dontrun{
qcChemFieldDupes()
qcChemFieldDupes(site = c("GRBA_L_DEAD0", "GRBA_L_JHNS0"), field.season = c("2018", "2019", "2020"))
}
}
