% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{qcChemFlags}
\alias{qcChemFlags}
\title{List all laboratory values that have an "Information," "Warning," or "Critical" flag.}
\usage{
qcChemFlags(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
List all laboratory values that have an "Information," "Warning," or "Critical" flag.
}
\examples{
\dontrun{
qcChemFlags()
qcChemFlags(site = c("GRBA_S_MILL1", "GRBA_S_PINE1", "GRBA_S_RDGE1"), field.season = c("2018", "2019", "2020"))
}
}
