% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{qcClosureErrorDiscrepancies}
\alias{qcClosureErrorDiscrepancies}
\title{Check the difference between closure errors calculated in R with those calculated in the Survey123 app}
\usage{
qcClosureErrorDiscrepancies(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Check the difference between closure errors calculated in R with those calculated in the Survey123 app
}
\examples{
\dontrun{
    qcStringSurveyElevations()
    qQcStringSurveyElevations(site = c("GRBA_L_BAKR0", "GRBA_L_JHNS0"), field.season = "2021")
}
}
