% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{qcElevationDiscrepancies}
\alias{qcElevationDiscrepancies}
\title{Check the difference between benchmark and water surface elevations calculated in R with those calculated in the Survey123 app}
\usage{
qcElevationDiscrepancies(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Check the difference between benchmark and water surface elevations calculated in R with those calculated in the Survey123 app
}
\examples{
\dontrun{
    qcStringSurveyElevations()
    qQcStringSurveyElevations(site = "GRBA_L_BAKR0", field.season = c("2019", "2021"))
}
}
