% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{qcLakeWqCleaned}
\alias{qcLakeWqCleaned}
\title{Intermediate step used to clean lake water quality data for stats and plotting functions.}
\usage{
qcLakeWqCleaned(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Limit data to primary visits and exclude data with "W" and "C" flags. Omit DO <110\% or <12 mg/L.
}
