% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{qcLakeWqSanity}
\alias{qcLakeWqSanity}
\title{Lake water quality sanity check}
\usage{
qcLakeWqSanity(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Perform sanity check and compile list of potentially incorrect or outlier water quality values.
}
