% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clarity-qc.R
\name{qcSecchiGTDepth}
\alias{qcSecchiGTDepth}
\title{List secchi depth measurements that are greater than the recorded lake depth}
\usage{
qcSecchiGTDepth(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_DEAD0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, VisitDate, FieldSeason, IsLakeDry, SurfaceCalm, OnBottom, DepthToBottom_m, SecchiDepth_m, VisitType, DPL.
}
\description{
List secchi depth measurements that are greater than the recorded lake depth
}
\examples{
\dontrun{
    qcSecchiGTDepth()
    qcSecchiGTDepth(site = "GRBA_L_BAKR0", field.season = "2019")
}
}
