% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{qcStringSurveyElevations}
\alias{qcStringSurveyElevations}
\title{Calculates mean and standard deviation of string survey lake level elevations for each year}
\usage{
qcStringSurveyElevations(park, site, field.season, sd_cutoff = NA)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{sd_cutoff}{Optional. If specified, only return benchmarks where standard deviation of final corrected elevations is greater than or equal to \code{sd_cutoff}.}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, VisitDate, FieldSeason, VisitType, MeanFinalElevation_ft, StDevFinalElevation_ft
}
\description{
Calculates mean and standard deviation of string survey lake level elevations for each year
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    qcStringSurveyElevations(conn)
    qQcStringSurveyElevations(conn, site = "GRBA_L_BAKR0", field.season = c("2016", "2017"))
    qcStringSurveyElevations(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
