% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{qcWqCleaned}
\alias{qcWqCleaned}
\title{Intermediate step used to clean water quality data for stats and plotting functions. Limit data to primary visits, and exclude data with "W" and "C" flags.}
\usage{
qcWqCleaned(park, site, field.season, wq.type)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{wq.type}{Either "stream" or "lake". Indicates whether to use stream or lake water quality data.}
}
\value{
A tibble
}
\description{
Intermediate step used to clean water quality data for stats and plotting functions. Limit data to primary visits, and exclude data with "W" and "C" flags.
}
