% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{qcWqCompletenessPlot}
\alias{qcWqCompletenessPlot}
\title{Plot percent completeness for each water quality parameter for each stream for each field season.}
\usage{
qcWqCompletenessPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot percent completeness for each water quality parameter for each stream for each field season.
}
\examples{
\dontrun{
    qcWqCompletenessPlot()
    qcWqCompletenessPlot(site = "GRBA_S_BAKR1", field.season = c("2018", "2019", "2020"))
}
}
