% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{qcWqGradesPlot}
\alias{qcWqGradesPlot}
\title{Plot the percentage of data rated at each grade level for each water quality parameter for each field season.}
\usage{
qcWqGradesPlot(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot the percentage of data rated at each grade level for each water quality parameter for each field season.
}
\examples{
\dontrun{
    qcWqGradesPlot()
    qcWqGradesPlot(site = "GRBA_S_BAKR1", field.season = c("2018", "2019", "2020"))
}
}
