% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{qcWqSanity}
\alias{qcWqSanity}
\title{Water quality sanity check}
\usage{
qcWqSanity(park, site, field.season, wq.type)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{wq.type}{Either "stream" or "lake". Indicates whether to use stream or lake water quality data.}
}
\value{
A tibble
}
\description{
Perform sanity check and compile list of potentially incorrect or outlier water quality values. This function is not exported; instead it is called by StreamQcWqSanity and LakeQcWqSanity
}
