% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadAGOL}
\alias{ReadAGOL}
\title{Read data from the Streams and Lakes AGOL feature layer}
\usage{
ReadAGOL(...)
}
\arguments{
\item{...}{}
}
\value{
A list of tibbles
}
\description{
Read data from the Streams and Lakes AGOL feature layer
}
