% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ReadAndFilterData}
\alias{ReadAndFilterData}
\title{Read Streams and Lakes data from database or .csv}
\usage{
ReadAndFilterData(park, site, field.season, data.name)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.name}{The name of the analysis view or the csv file containing the data. E.g. "CalibrationDO", "DischargeVolumetric". See details for full list of data name options.}
}
\value{
A tibble of filtered data.
}
\description{
Read Streams and Lakes data from database or .csv
}
\details{
\code{data.name} options are: Site, Visit, BMIVisit, BMIMetrics, BMISpecies, Channel, Chemistry, Clarity, WaterQualityDO, WaterQualitypH, WaterQualitySpCond, WaterQualityTemperature, WQStreamXSection, CalibrationDO, CalibrationpH, CalibrationSpCond, TimeseriesDOmgl, TimeseriesDOpct, TimeseriespH, TimeseriesSpCond, TimeseriesTemperature, TimeseriesDischarge, TimeseriesWaterLevel
}
