% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{SaveDataToCsv}
\alias{SaveDataToCsv}
\title{Save Streams and Lakes analysis views as a set of .csv files}
\usage{
SaveDataToCsv(
  conn,
  dest.folder,
  create.folders = FALSE,
  overwrite = FALSE,
  aquarius = TRUE,
  calculated = TRUE
)
}
\arguments{
\item{conn}{A database connection pool object generated from a call to \code{OpenDatabaseConnection()}.}

\item{dest.folder}{The folder in which to save the .csv files.}

\item{create.folders}{Should \code{dest.folder} be created automatically if it doesn't exist? Defaults to \code{FALSE}.}

\item{overwrite}{Should existing data be automatically overwritten? Defaults to \code{FALSE}.}

\item{aquarius}{Include Aquarius data?}

\item{calculated}{Include calculated data (median stream wq, median lake wq, and lake surface elevation)?}
}
\value{
None.
}
\description{
Save Streams and Lakes analysis views as a set of .csv files
}
\examples{
\dontrun{
conn <- OpenDatabaseConnection()
SaveDataToCsv(conn, "C:/Users/myusername/Documents/R/streamsandlakes-data", TRUE, TRUE)
CloseDatabaseConnection(conn)
}
}
