% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{StringSurveyElevation}
\alias{StringSurveyElevation}
\title{Calculates mean elevations for each survey point type in a survey (string)}
\usage{
StringSurveyElevation(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Calculates mean elevations for each survey point type in a survey (string)
}
\examples{
\dontrun{
    StringSurveyElevation()
    StringSurveyElevation(site = "GRBA_L_BAKR0", field.season = "2016")
}
}
