% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wq-qc.R
\name{WqPlotDepthProfile}
\alias{WqPlotDepthProfile}
\title{Generate lake depth profile plots.}
\usage{
WqPlotDepthProfile(
  param,
  units,
  park,
  site,
  field.season,
  include.title = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{param}{The water quality parameter to plot. One of "pH", "DO", "SpCond", or "Temperature".}

\item{units}{Units of dissolved oxygen. Either "mg/L" or "\%". Ignored if \code{param != "DO"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{include.title}{Include plot title? Defaults to true.}

\item{plotly}{Return an interactive plotly object instead of a ggplot object? Defaults to false.}
}
\value{
ggplot or plotly object
}
\description{
Generate lake depth profile plots.
}
