% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{ChemANC}
\alias{ChemANC}
\title{Calculate acid neutralizing capacity (ANC) from alkalinity (ALK2)}
\usage{
ChemANC(conn, path.to.data, park, site, field.season, data.source = "database")
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live Streams and Lakes database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, FieldSeason, SampleFrame, VisitDate, VisitType, SampleCollectionMethod, Characteristic, CharacteristicLabel, LabValue, ReportingGroup, SampleType, Flag, FlagNote, DPL, Unit
}
\description{
Calculate acid neutralizing capacity (ANC) from alkalinity (ALK2)
}
\examples{
\dontrun{
conn <- OpenDatabaseConnection()
ChemANC(conn)
ChemANC(conn, site = "GRBA_L_DEAD0", field.season = "2018")
CloseDatabaseConnection(conn)
}
}
