% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetAquariusColSpec}
\alias{GetAquariusColSpec}
\title{Get column specifications for Aquarius database}
\usage{
GetAquariusColSpec()
}
\value{
A list of column specifications for each table of Aquarius data.
}
\description{
Get column specifications for Aquarius database
}
