% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{GetColSpec}
\alias{GetColSpec}
\title{Get column specifications for SQL database}
\usage{
GetColSpec()
}
\value{
A list of column specifications for each table of SQL data.
}
\description{
Get column specifications for SQL database
}
