% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{PlotBenchmarkElevation}
\alias{PlotBenchmarkElevation}
\title{Plot benchmark elevations over time}
\usage{
PlotBenchmarkElevation(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database",
  include.title = TRUE,
  plotly = FALSE
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live Streams and Lakes database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{include.title}{Include plot title? Defaults to true.}

\item{plotly}{Return an interactive plotly object instead of a ggplot object? Defaults to false.}
}
\value{
A ggplot or plotly object
}
\description{
Plot benchmark elevations over time
}
\examples{
\dontrun{
conn <- OpenDatabaseConnection()
PlotBenchmarkElevation(conn)
PlotBenchmarkElevation(conn, site = "GRBA_L_DEAD0", plotly = TRUE)
CloseDatabaseConnection(conn)
}
}
