% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{PlotStringComparisons}
\alias{PlotStringComparisons}
\title{Plot lake levels determined by all benchmarks where string method was used}
\usage{
PlotStringComparisons(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR1".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A ggplot object
}
\description{
Plot lake levels determined by all benchmarks where string method was used
}
\examples{
\dontrun{
PlotStringComparisons()
PlotStringComparisons(site = "GRBA_L_DEAD0")
}
}
