% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chem-qc.R
\name{qcChemML}
\alias{qcChemML}
\title{List all routine laboratory values that are less than or equal to the minimum level of quantitation (ML) for that analyte.}
\usage{
qcChemML(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns SampleFrame, SiteCode, SiteName, FieldSeason, VisitDate, Characteristic, CharacteristicLabel, Unit, LabValue, ML, MLFlag.
}
\description{
List all routine laboratory values that are less than or equal to the minimum level of quantitation (ML) for that analyte.
}
\examples{
\dontrun{
qcChemML()
qcChemML(site = c("GRBA_L_DEAD0", "GRBA_L_JHNS0"), field.season = c("2018", "2019", "2020"))
}
}
