% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{qcDPLCheck}
\alias{qcDPLCheck}
\title{Return list of site visits that have any data categorized as "Raw" or "Provisional"}
\usage{
qcDPLCheck(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns SiteCode, SiteName, VisitDate, FieldSeason, SampleFrame, VisitType, Visit.DPL, Chem.DPL, BMI.DPL, Channel.DPL, Clarity.DPL, LakeSurvey.DPL, LakeString.DPL, Xsection.DPL, TempC.DPL, pH.DPL, SpCond.DPL, DO.DPL
}
\description{
Return list of site visits that have any data categorized as "Raw" or "Provisional"
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    qcDPLCheck(conn)
    qcDPLCheck(conn, site = "GRBA_L_JHNS0", field.season = c("2018", "2019", "2020"))
    CloseDatabaseConnection(conn)
}
}
