% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{qcNoAnnualVisit}
\alias{qcNoAnnualVisit}
\title{Return list of streams and lakes that were not visited for annual monitoring during a field season}
\usage{
qcNoAnnualVisit(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, SampleFrame, FieldSeason, VisitDate
}
\description{
Return list of streams and lakes that were not visited for annual monitoring during a field season
}
\examples{
\dontrun{
    qcNoAnnualVisit()
    qcNoAnnualVisit(site = "GRBA_L_DEAD0", field.season = c("2012", "2013", "2014", "2015"))
}
}
