% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{qcStringSurveyHeights}
\alias{qcStringSurveyHeights}
\title{Calculates mean and standard deviation of string survey heights for each benchmark}
\usage{
qcStringSurveyHeights(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  data.source = "database",
  sd_cutoff = NA
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{data.source}{Character string indicating whether to access data in the live Streams and Lakes database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}

\item{sd_cutoff}{Optional. If specified, only return benchmarks where standard deviation of final corrected elevations is greater than or equal to \code{sd_cutoff}.}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, VisitDate, FieldSeason, VisitType, MeanHeight_ft, StDevHeight_ft
}
\description{
Calculates mean and standard deviation of string survey heights for each benchmark
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    QcStringSurveyHeights(conn)
    QcStringSurveyHeights(conn, site = "GRBA_L_BAKR0", field.season = c("2016", "2017"))
    QcStringSurveyHeights(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
