% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{qcWqGradesPlot}
\alias{qcWqGradesPlot}
\title{Plot the percentage of data rated at each grade level for each water quality parameter for each field season.}
\usage{
qcWqGradesPlot(
  conn,
  path.to.data,
  park,
  site,
  field.season,
  parameter,
  data.source = "database"
)
}
\arguments{
\item{conn}{Database connection generated from call to \code{OpenDatabaseConnection()}. Ignored if \code{data.source} is \code{"local"}.}

\item{path.to.data}{The directory containing the csv data exports generated from \code{SaveDataToCsv()}. Ignored if \code{data.source} is \code{"database"}.}

\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}

\item{parameter}{Optional. Water quality parameter to filter on.}

\item{data.source}{Character string indicating whether to access data in the live Streams and Lakes database (\code{"database"}, default) or to use data saved locally (\code{"local"}). In order to access the most up-to-date data, it is recommended that you select \code{"database"} unless you are working offline or your code will be shared with someone who doesn't have access to the database.}
}
\value{
A ggplot object
}
\description{
Plot the percentage of data rated at each grade level for each water quality parameter for each field season.
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    qcWqGradesPlot(conn)
    qcWqGradesPlot(conn, site = "GRBA_S_BAKR1", field.season = c("2018", "2019", "2020"))
    CloseDatabaseConnection(conn)
}
}
