% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/levels-qc.R
\name{SurveyPointElevation}
\alias{SurveyPointElevation}
\title{Calculates mean elevations for each survey point type in a survey}
\usage{
SurveyPointElevation(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble with columns Park, SiteShort, SiteCode, SiteName, VisitDate, FieldSeason,VisitType, DPL, SurveyPoint, Benchmark, FinalCorrectedElevation_ft.
}
\description{
Calculates mean elevations for each survey point type in a survey
}
\examples{
\dontrun{
    conn <- OpenDatabaseConnection()
    SurveyPointElevation(conn)
    SurveyPointElevation(conn, site = "GRBA_L_BAKR0", field.season = "2019")
    SurveyPointElevation(path.to.data = "path/to/data", data.source = "local")
    CloseDatabaseConnection(conn)
}
}
