% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efficiency-qc.R
\name{WqDailyMean}
\alias{WqDailyMean}
\title{Return summary of daily mean values (daily median values for pH) and grades for water quality parameters at streams.}
\usage{
WqDailyMean(park, site, field.season)
}
\arguments{
\item{park}{Optional. Four-letter park code to filter on, e.g. "GRBA".}

\item{site}{Optional. Site code to filter on, e.g. "GRBA_L_BAKR0".}

\item{field.season}{Optional. Field season name to filter on, e.g. "2019".}
}
\value{
A tibble
}
\description{
Return summary of daily mean values (daily median values for pH) and grades for water quality parameters at streams.
}
\examples{
\dontrun{
    WqDailyMean()
    WqDailyMean(site = "GRBA_S_BAKR1", field.season = c("2018", "2019", "2020"))
}
}
